+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/FLTCamZoomTests.swift
@@ -9,16 +9,16 @@ import XCTest
 
 // Import Objectice-C part of the implementation when SwiftPM is used.
 #if canImport(camera_avfoundation_objc)
-  @testable import camera_avfoundation_objc
+  import camera_avfoundation_objc
 #endif
 
 final class FLTCamZoomTests: XCTestCase {
-  private func createCamera() -> (FLTCam, MockCaptureDevice) {
+  private func createCamera() -> (Camera, MockCaptureDevice) {
     let mockDevice = MockCaptureDevice()
 
     let configuration = CameraTestUtils.createTestCameraConfiguration()
-    configuration.captureDeviceFactory = { mockDevice }
-    let camera = FLTCam(configuration: configuration, error: nil)
+    configuration.captureDeviceFactory = { _ in mockDevice }
+    let camera = CameraTestUtils.createTestCamera(configuration)
 
     return (camera, mockDevice)
   }